Password form rows moved to dedicate reusable component
